fix(fe): move confirmNavigation to utils file
